Done from addButton for personne and entroprise both for CLIUENT and FOURNISSEUR

+++ gieBackend/src/main/java/org/gieback/Service/EntrepriseService.java
@@ -35,7 +35,7 @@ public class EntrepriseService implements IEntrepriseService{
     }
 
     @Override
-    public void deleteByid(int id) {
+    public void deleteByid(Long id) {
         entrepriseDao.deleteByid(id);
 
     }
