+++ gieFrontend/src/main/java/org/giefront/Service/PersonneService.java
@@ -160,7 +160,7 @@ public class PersonneService implements IService{
         }
     }
 
-    public void deletePersonne(int id) throws IOException {
+    public void deletePersonne(Long id) throws IOException {
         Request request = new Request.Builder().url("http://localhost:9998/personne/DeletePersonne/"+id).delete().build();
         try (Response response = okHttpClient.newCall(request).execute()) {
             if (!response.isSuccessful()) {
@@ -217,6 +217,8 @@ public class PersonneService implements IService{
             throw new IOException("Erreur lors de l'exécution de la requête HTTP", e);
         }
     }
+
+
 
 
 }
